fix(vfs): fallocate should only extend file, not shrink
    According to Linux fallocate(2) semantics, mode=0 should only
    extend the file size when offset+len exceeds the current file
    size. It should never shrink the file.

+++ kernel/src/filesystem/vfs/syscall/sys_fallocate.rs
@@ -75,7 +75,7 @@ impl Syscall for SysFallocateHandle {
 
             let new_size = offset.saturating_add(len);
 
-            let current_size = file.inode().metadata()?.size;
+            let current_size = file.inode().metadata()?.size as usize;
             if new_size > current_size {
                 let current_pcb = ProcessManager::current_pcb();
                 let fsize_limit = current_pcb.get_rlimit(RLimitID::Fsize);
